requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
python-multipart
jinja2
huggingface_hub